refactor - move ItemEnqueuePositionCalculator to top-level per review.
